commit po dlhej dobe1

+++ src/main/java/com/example/kniznica/Controller.java
@@ -6,7 +6,12 @@ import com.example.kniznica.repositories.bookRepository;
 import com.example.kniznica.services.booksService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
+/**
+ * 
+ * 
+ * 
+ * 
+ */
 
 @RestController
 //@RequestMapping("api/")
@@ -20,43 +25,42 @@ public class Controller {
      bookRepository bookRepository;
 
     @GetMapping("getbooks")
-
     public List<books> getAllBook() {
-
+     
          return booksService.getAllBooks();
  
     }
- 
-    @GetMapping("getbooks/{id}")
 
+    @GetMapping("getbook/{id}")
     public books getBook(@PathVariable Long id) {
 
          return booksService.getBook(id);
  
     }
 
-    @PostMapping("addbook")
-    public void saveBooks(@RequestBody books books) {
-           booksService.saveBooks(books);
+    /**za du findByName */
+
+    @PostMapping("savebook")
+    public void saveBook(@RequestBody books books) {
+           booksService.saveBook(books);
          
     }
 
     @RequestMapping(method = RequestMethod.PUT, value="addbooks/{id}")
-
     public void updateBooks(@PathVariable Long id, @RequestBody books books) {
       booksService.updateBooks(id,books);
           
-    }
-
+    } 
+/**
     @DeleteMapping(value = "deletebook/{id}")
-
     public void deleteBooks(@PathVariable Long id) {
+     if(!booksService.deleteBooks(id))throw new BookNotfoundException();
       booksService.deleteBooks(id);
           
     }
+ */    
 
     @RequestMapping("**")
-
     public String nothing() {
 
           return "You have entered an incorrect URL !!!";
